refactor: ApiResponse to MakeJsonResponse
Rename ApiResponse, to best semantic name, and clean up code on new trait
