+++ app/Http/Controllers/Admin/User/UserController.php
@@ -7,32 +7,44 @@ namespace App\Http\Controllers\Admin\User;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\User\UpdateRequest;
 use App\Repositories\User\UserRepositoryInterface;
-use App\Traits\Responses\ApiResponse;
+use App\Traits\Responses\MakeJsonResponse;
 use Illuminate\View\View;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class UserController extends Controller
+final class UserController extends Controller
 {
 
-    use ApiResponse;
+    use MakeJsonResponse;
 
     private UserRepositoryInterface $repository;
 
+
+    /**
+     * New Admin/UserController instance
+     *
+     * @param UserRepositoryInterface $repository
+     */
     public function __construct(UserRepositoryInterface $repository)
     {
         $this->repository = $repository;
     }
 
+
     /**
      * Display a listing of the resource.
+     *
+     * @param Request $request
+     * @return JsonResponse|View
      */
     public function index(Request $request): JsonResponse|View
     {
         if ($request->wantsJson()) {
-            return $this->showAll(collection: $this->repository->all($request->all()));
+
+            return $this->successResponse(data: $this->repository->all($request->all()));
         } else {
+
             return view('admin.users.index');
         }
     }
@@ -40,6 +52,9 @@ class UserController extends Controller
 
     /**
      * Show the form for editing the specified resource.
+     *
+     * @param integer $id
+     * @return View
      */
     public function edit(int $id): View
     {
@@ -49,15 +64,25 @@ class UserController extends Controller
         ]);
     }
 
+
     /**
      * Update the specified resource in storage.
+     *
+     * @param UpdateRequest $request
+     * @param integer $id
+     * @return JsonResponse
      */
     public function update(UpdateRequest $request, int $id): JsonResponse
     {
         if ($this->repository->update($request->validated(), $id)) {
+
             return $this->showMessage(message: trans('server.record_updated'));
         } else {
-            return $this->errorResponseWithBag(collection: ['server' => [trans('server.internal_error')]], code: Response::HTTP_INTERNAL_SERVER_ERROR);
+
+            return $this->errorResponseWithBag(
+                collection: ['server' => [trans('server.internal_error')]],
+                code: Response::HTTP_INTERNAL_SERVER_ERROR
+            );
         }
     }
 }
